issue #35 - checks in seperate file
